+++ src/containers/(member)/join/JoinForm.tsx
@@ -2,16 +2,16 @@
 
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import FirstBtnArea from '@/containers/join/FirstBtnArea'
-import IdInput from '@/containers/join/IdInput'
-import NicknameInput from '@/containers/join/NicknameInput'
-import PhoneVerification from '@/containers/join/PhoneVerification'
-import ProfileImgArea from '@/containers/join/ProfileImgArea'
-import Pw2Input from '@/containers/join/Pw2Input'
-import PwInput from '@/containers/join/PwInput'
-import SecondBtnArea from '@/containers/join/SecondBtnArea'
-import SelectCategory from '@/containers/join/SelectCategory'
-import { useFirstStore, useSecondStore } from '@/containers/join/store'
+import FirstBtnArea from '@/containers/(member)/join/FirstBtnArea'
+import IdInput from '@/containers/(member)/join/IdInput'
+import NicknameInput from '@/containers/(member)/join/NicknameInput'
+import PhoneVerification from '@/containers/(member)/join/PhoneVerification'
+import ProfileImgArea from '@/containers/(member)/join/ProfileImgArea'
+import Pw2Input from '@/containers/(member)/join/Pw2Input'
+import PwInput from '@/containers/(member)/join/PwInput'
+import SecondBtnArea from '@/containers/(member)/join/SecondBtnArea'
+import SelectCategory from '@/containers/(member)/join/SelectCategory'
+import { useFirstStore, useSecondStore } from '@/containers/(member)/join/store'
 import { join } from '@/utils/api/memberFunctions'
 import { uploadImage } from '@/utils/uploadImage'
 import DuckOne from '@/../public/svgs/duck/duckone.svg'
